+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixes
+
+- refactor code to create a lot less policies in order to improve performance
+
 ## 2023-07-17 0.1.0
 
 ### Features